cli-app/requirements.txt
pdfplumber 
python-dotenv
google-auth 
google-auth-oauthlib 
google-api-python-client 
python-dotenv